Generalized the cluster class
